fix: increment the maximum size of offsets in regexp intructions
The offset associated with jump and split instructions in regular expressions was represented with a i32, which limited maximum length of certain jumps.

+++ yara-x/src/re/instr.rs
@@ -74,7 +74,7 @@ pub type NumAlt = u8;
 
 /// Offset for jump and split instructions. The offset is always relative to
 /// the address where the instruction starts.
-pub type Offset = i16;
+pub type Offset = i32;
 
 /// Instructions supported by the Pike VM.
 pub enum Instr<'a> {
